Clean up ChasmClassWriter
